refactor(service): 불필요한 인터페이스 삭제
