+++ src/main/java/subway/service/StationService.java
@@ -1,11 +1,24 @@
 package subway.service;
 
+import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import subway.domain.Station;
 import subway.dto.StationCreateRequest;
 import subway.dto.StationResponse;
+import subway.repository.StationRepository;
 
-public interface StationService {
+@Service
+public class StationService {
+    private final StationRepository stationRepository;
 
-    @Transactional(readOnly = true)
-    StationResponse createStation(StationCreateRequest request);
+    public StationService(final StationRepository stationRepository) {
+        this.stationRepository = stationRepository;
+    }
+
+    @Transactional
+    public StationResponse createStation(final StationCreateRequest request) {
+        final Station station = stationRepository.save(new Station(request.getName()));
+        return StationResponse.from(station);
+    }
 }
+
